third_party/WebKit/LayoutTests/fast/css/fontloader-multiple-families-expected.txt

Test load events for fonts.loadFont() with multiple font families.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.fontfaces.length is 2
PASS event.fontfaces[0].status is "loaded"
PASS event.fontfaces[1].status is "loaded"
PASS document.fonts.check('10px TestFont1') is true
PASS document.fonts.check('10px TestFont2') is true
PASS document.fonts.check('10px TestFont1, TestFont2') is true
PASS firedEvents is ['loading', 'loadingdone']
PASS successfullyParsed is true

TEST COMPLETE
